MAINT move dataset get logic back into tests file
